Internal refactoring to clarify separation of concerns
This is a minor internal reorganization to clarify how the IO
requirements interact with the control flow requirements. The former
is provided by a free monad transformer (HppT) that provides specific
operations, while the latter are provided by stock definitions from
the `transformers`.

To clarify this split, I added a `dischargeHppCaps` function that
provides the needed plumbing using `StateT` and `ExceptT`.